# SequenceTokenParser: Use `ROS<T>.CommonPrefixLength()`.