Gives a nice speed-up to Parser.String().

Benchmark on my machine (`StringBench.String`):

| Method  | Mean      | Error     | StdDev    | Gen0     | Gen1     | Gen2     | Allocated |
|-------- |----------:|----------:|----------:|---------:|---------:|---------:|----------:|
| main `String`    | 15.320 us | 0.1890 us | 0.1770 us |        - |        - |        - |      40 B |
| **pr** `String`     |  **1.912 us** | 0.0044 us | 0.0042 us |        - |        - |        - |      40 B |