@benjamin-hodgson 2025-06-27
Shipped in v3.5.0. Thanks!